+++ lang/java/reef-runtime-spark/src/test/java/org/apache/reef/runtime/spark/job/ReefOnSparkDriver.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.reef.runtime.spark;
+package org.apache.reef.runtime.spark.job;
 
 import org.apache.reef.driver.evaluator.AllocatedEvaluator;
 import org.apache.reef.driver.evaluator.EvaluatorRequestor;
